requirements.txt
pytest-playwright
py
pytest
pytest-bdd
pytest-html
pytest-bdd-html
